move password reset routes out of the public api, into session.
adjust the email template
