+++ src/api/routes/public/public-user.php
@@ -10,8 +10,7 @@ use Slim\Http\Response;
 $app->group('/public/user', function () {
     $this->post('/login', 'userLogin');
     $this->post('/login/', 'userLogin');
-    $this->post('/password/reset', 'userPasswordReset');
-    $this->post('/password/reset/', 'userPasswordReset');
+
 });
 
 function userLogin(Request $request, Response $response, array $args)
@@ -31,25 +30,3 @@ function userLogin(Request $request, Response $response, array $args)
     return $response->withStatus(404);
 }
 
-function userPasswordReset(Request $request, Response $response, array $args)
-{
-    $logger = LoggerUtils::getAppLogger();
-    $body = json_decode($request->getBody());
-    $userName = strtolower(trim($body->userName));
-    if (!empty($userName)) {
-        $user = UserQuery::create()->findOneByUserName($userName);
-        if (!empty($user) && !empty($user->getEmail())) {
-            $token = new Token();
-            $token->build("password", $user->getId());
-            $token->save();
-            $email = new ResetPasswordTokenEmail($user, $token->getToken());
-            if (!$email->send()) {
-                LoggerUtils::getAppLogger()->error($email->getError());
-            }
-            return $response->withStatus(200);
-        } else {
-            return $response->withStatus(404, gettext("User") . " [" . $userName . "] ". gettext("no found or user without an email"));
-        }
-    }
-    return $response->withStatus(400, gettext("UserName not set"));
-}
